Refactor: 로그인 시 프로필 저장
