+++ src/main/java/supernova/whokie/global/config/WebMvcConfig.java
@@ -33,10 +33,8 @@ public class WebMvcConfig implements WebMvcConfigurer {
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         // Todo 경로 수정 해야함
-        registry.addInterceptor(jwtInterceptor())
-            .addPathPatterns("/api/user/mypage")
-            .addPathPatterns("/api/profile/**")
-        ;
+        /*registry.addInterceptor(jwtInterceptor())
+            .addPathPatterns("/api/**");*/
     }
 
     @Override
