+++ src/main/java/supernova/whokie/profile/Profile.java
@@ -6,6 +6,8 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToOne;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -28,8 +30,10 @@ public class Profile {
     @JoinColumn(name = "user_id")
     private Users users;
 
+    @NotNull
     private Integer todayVisited;
 
+    @NotNull
     private Integer totalVisited;
 
     private String description;
