--- src/main/java/supernova/whokie/user/infrastructure/apiCaller/dto/Partner.java
@@ -1,5 +1,0 @@
-package supernova.whokie.user.infrastructure.apiCaller.dto;
-
-public record Partner(String uuid) {
-
-}
