+++ src/main/java/supernova/whokie/user/infrastructure/apiCaller/dto/UserInfoResponse.java
@@ -4,6 +4,6 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record UserInfoResponse(Long id, KakaoAccount kakaoAccount, Partner forPartner) {
+public record UserInfoResponse(Long id, KakaoAccount kakaoAccount) {
 
 }
