+++ src/main/java/supernova/whokie/user/infrastructure/repository/UsersRepository.java
@@ -1,4 +1,4 @@
-package supernova.whokie.user.repository;
+package supernova.whokie.user.infrastructure.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import supernova.whokie.user.Users;
