+++ src/main/java/supernova/whokie/user/service/UserService.java
@@ -6,13 +6,14 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import supernova.whokie.global.auth.JwtProvider;
 import supernova.whokie.global.exception.EntityNotFoundException;
+import supernova.whokie.profile.Profile;
+import supernova.whokie.profile.infrastructure.ProfileRepository;
 import supernova.whokie.user.Gender;
 import supernova.whokie.user.Role;
 import supernova.whokie.user.Users;
 import supernova.whokie.user.infrastructure.apiCaller.dto.KakaoAccount;
-import supernova.whokie.user.infrastructure.apiCaller.dto.Partner;
 import supernova.whokie.user.infrastructure.apiCaller.dto.UserInfoResponse;
-import supernova.whokie.user.repository.UserRepository;
+import supernova.whokie.user.infrastructure.repository.UserRepository;
 import supernova.whokie.user.infrastructure.apiCaller.UserApiCaller;
 import supernova.whokie.user.service.dto.UserModel;
 
@@ -21,6 +22,7 @@ import supernova.whokie.user.service.dto.UserModel;
 public class UserService {
 
     private final UserRepository userRepository;
+    private final ProfileRepository profileRepository;
     private final JwtProvider jwtProvider;
     private final UserApiCaller userApiCaller;
 
@@ -32,20 +34,32 @@ public class UserService {
     public String register(String code) {
         UserInfoResponse userInfoResponse = userApiCaller.extractUserInfo(code);
         KakaoAccount kakaoAccount = userInfoResponse.kakaoAccount();
-        Partner partner = userInfoResponse.forPartner();
+
         Users user = userRepository.findByEmail(kakaoAccount.email())
-            .orElseGet(() -> userRepository.save(
-                Users.builder()
-                    .name(kakaoAccount.name())
-                    .email(kakaoAccount.email())
-                    .point(0)
-                    .age(LocalDate.now().getYear() - Integer.parseInt(kakaoAccount.birthYear()))
-                    .gender(Gender.fromString(kakaoAccount.gender()))
-                    .imageUrl(kakaoAccount.profile().profileImageUrl())
-                    .role(Role.USER)
-                    .kakaoId(userInfoResponse.id())
-                    .build()
-            ));
+            .orElseGet(() -> {
+                Users newUser = userRepository.save(
+                    Users.builder()
+                        .name(kakaoAccount.name())
+                        .email(kakaoAccount.email())
+                        .point(0)
+                        .age(LocalDate.now().getYear() - Integer.parseInt(kakaoAccount.birthYear()))
+                        .gender(Gender.fromString(kakaoAccount.gender()))
+                        .imageUrl(kakaoAccount.profile().profileImageUrl())
+                        .role(Role.USER)
+                        .kakaoId(userInfoResponse.id())
+                        .build()
+                );
+
+                Profile profile = Profile.builder()
+                    .users(newUser)
+                    .todayVisited(0)
+                    .totalVisited(0)
+                    .backgroundImageUrl(kakaoAccount.profile().profileImageUrl())
+                    .build();
+
+                profileRepository.save(profile);
+                return newUser;
+            });
 
         String token = jwtProvider.createToken(user.getId(), user.getRole());
         System.out.println(token);
