+++ src/test/java/supernova/whokie/friend/repository/FriendRepositoryTest.java
@@ -10,7 +10,7 @@ import supernova.whokie.friend.infrastructure.repository.FriendRepository;
 import supernova.whokie.user.Gender;
 import supernova.whokie.user.Role;
 import supernova.whokie.user.Users;
-import supernova.whokie.user.repository.UserRepository;
+import supernova.whokie.user.infrastructure.repository.UserRepository;
 
 import java.util.List;
