+++ src/test/java/supernova/whokie/friend/service/FriendServiceTest.java
@@ -21,7 +21,7 @@ import supernova.whokie.global.auth.JwtProvider;
 import supernova.whokie.user.Gender;
 import supernova.whokie.user.Role;
 import supernova.whokie.user.Users;
-import supernova.whokie.user.repository.UserRepository;
+import supernova.whokie.user.infrastructure.repository.UserRepository;
 
 import java.util.ArrayList;
 import java.util.List;
