+++ src/test/java/supernova/whokie/user/controller/UserControllerTest.java
@@ -22,7 +22,7 @@ import supernova.whokie.global.auth.JwtProvider;
 import supernova.whokie.user.Gender;
 import supernova.whokie.user.Role;
 import supernova.whokie.user.Users;
-import supernova.whokie.user.repository.UserRepository;
+import supernova.whokie.user.infrastructure.repository.UserRepository;
 
 @SpringBootTest
 @AutoConfigureMockMvc
