+++ src/test/java/supernova/whokie/user/repository/UserRepositoryTest.java
@@ -8,6 +8,7 @@ import org.springframework.test.annotation.DirtiesContext;
 import supernova.whokie.user.Users;
 
 import java.util.List;
+import supernova.whokie.user.infrastructure.repository.UserRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
